@@ -30,12 +30,13 @@
     "isomorphic-fetch": "^2.2.1",
     "jiminy": "^0.1.1",
     "json-loader": "https://registry.npmjs.org/json-loader/-/json-loader-0.5.4.tgz",
+    "leaflet": "^1.0.3",
     "lodash": "https://registry.npmjs.org/lodash/-/lodash-4.17.4.tgz",
     "lusca": "^1.4.1",
     "moveto": "^1.6.1",
-    "next": "^2.0.0",
+    "next": "^2.4.3",
     "next-redux-wrapper": "https://registry.npmjs.org/next-redux-wrapper/-/next-redux-wrapper-1.1.2.tgz",
-    "next-routes": "^1.0.24",
+    "next-routes": "^1.0.27",
     "node-sass": "^4.5.0",
     "nodemailer": "^2.7.0",
     "normalize.css": "^7.0.0",
@@ -73,7 +74,7 @@
     "three": "^0.83.0",
     "url-loader": "^0.5.8",
     "uuid": "^3.0.1",
-    "vega": "^3.0.0-beta.31",
+    "vega": "^2.6.0",
     "whatwg-fetch": "https://registry.npmjs.org/whatwg-fetch/-/whatwg-fetch-2.0.2.tgz"
   },
   "devDependencies": {
