@@ -1,4 +1,7 @@
 import React from 'react';
+import withRedux from 'next-redux-wrapper';
+import { initStore } from 'store';
+import { Link } from 'routes';
 
 // Components
 import Title from 'components/ui/Title';
@@ -6,8 +9,7 @@ import Button from 'components/ui/Button';
 import DatasetWidgetChart from 'components/app/explore/DatasetWidgetChart';
 import DatasetLayerChart from 'components/app/explore/DatasetLayerChart';
 import DatasetPlaceholderChart from 'components/app/explore/DatasetPlaceholderChart';
-import { Link } from 'routes';
-
+import { toggleDatasetActive, setUrlParams, setDatasetsHidden } from 'redactions/explore';
 
 class DatasetWidget extends React.Component {
 
@@ -47,40 +49,6 @@ class DatasetWidget extends React.Component {
    * - getDescription
    * - getButton
   */
-  hideLayer(dataset) {
-    let newLayersHidden = this.props.layersHidden.slice();
-    this.props.layersHidden.includes(dataset) ?
-      newLayersHidden = this.props.layersHidden.filter(l => l !== dataset) :
-      newLayersHidden.push(dataset);
-
-    this.props.setDatasetsHidden(newLayersHidden);
-  }
-
-  getDescription(_text) {
-    let text = _text;
-    if (typeof text === 'string' && text.length > 70) {
-      text = text.replace(/^(.{70}[^\s]*).*/, '$1');
-      return `${text}...`;
-    }
-    return text;
-  }
-
-  getWidgetOrLayer() {
-    if (this.state.hasWidget) {
-      return {
-        ...this.state.widget.attributes,
-        ...{ id: this.state.widget.id }
-      };
-    }
-    if (this.state.hasLayer) {
-      return {
-        ...this.state.layer.attributes,
-        ...{ id: this.state.layer.id }
-      };
-    }
-    return null;
-  }
-
   getButton() {
     const { active, layer } = this.state;
     const buttonText = (active) ? 'Active' : 'Add to map';
@@ -112,6 +80,42 @@ class DatasetWidget extends React.Component {
     );
   }
 
+  getWidgetOrLayer() {
+    if (this.state.hasWidget) {
+      return {
+        ...this.state.widget.attributes,
+        id: this.state.widget.id
+      };
+    }
+    if (this.state.hasLayer) {
+      return {
+        ...this.state.layer.attributes,
+        id: this.state.layer.id
+      };
+    }
+    return null;
+  }
+
+  getDescription(_text) {
+    let text = _text;
+    if (typeof text === 'string' && text.length > 70) {
+      text = text.replace(/^(.{70}[^\s]*).*/, '$1');
+      return `${text}...`;
+    }
+    return text;
+  }
+
+  hideLayer(dataset) {
+    let newLayersHidden = this.props.layersHidden.slice();
+    if (this.props.layersHidden.includes(dataset)) {
+      newLayersHidden = this.props.layersHidden.filter(l => l !== dataset);
+    } else {
+      newLayersHidden.push(dataset);
+    }
+
+    this.props.setDatasetsHidden(newLayersHidden);
+  }
+
   /**
    * UI EVENTS
    * - triggerToggleLayer (e)
@@ -119,7 +123,9 @@ class DatasetWidget extends React.Component {
   triggerToggleLayer() {
     const { dataset } = this.state;
     this.props.toggleDatasetActive(dataset.id);
-    this.props.layersHidden.includes(dataset.id) && this.hideLayer(dataset.id);
+    if (this.props.layersHidden.includes(dataset.id)) {
+      this.hideLayer(dataset.id);
+    }
   }
 
   render() {
@@ -187,4 +193,16 @@ DatasetWidget.propTypes = {
   setDatasetsHidden: React.PropTypes.func
 };
 
-export default DatasetWidget;
+const mapStateToProps = state => ({
+  layersHidden: state.explore.datasets.hidden
+});
+
+const mapDispatchToProps = dispatch => ({
+  toggleDatasetActive: (id) => {
+    dispatch(toggleDatasetActive(id));
+    dispatch(setUrlParams());
+  },
+  setDatasetsHidden: (id) => { dispatch(setDatasetsHidden(id)); }
+});
+
+export default withRedux(initStore, mapStateToProps, mapDispatchToProps)(DatasetWidget);
