@@ -42,7 +42,8 @@ module.exports = {
         'process.env.TEMP_TOKEN': JSON.stringify(process.env.TEMP_TOKEN),
         'process.env.BACKOFFICE_API_URL': JSON.stringify(process.env.BACKOFFICE_API_URL),
         'process.env.WRI_API_URL': JSON.stringify(process.env.WRI_API_URL),
-        'process.env.CMS_API_URL': JSON.stringify(process.env.CMS_API_URL)
+        'process.env.CMS_API_URL': JSON.stringify(process.env.CMS_API_URL),
+        'process.env.BASEMAP_TILE_URL': JSON.stringify(process.env.BASEMAP_TILE_URL)
       })
     );
     return config;
