@@ -6,4 +6,4 @@ export { default as partners } from './partners';
 export { default as partnerDetail } from './partnerDetail';
 export { default as tooltip } from './tooltip';
 export { default as staticPages } from './static_pages';
-// export { modalReducer as modal } from 'rw-components';
+export { default as modal } from './modal';
