@@ -16,7 +16,7 @@ const initialState = {
   loading: false
 };
 
-export default function modalReducer(state = initialState, action) {
+export default function (state = initialState, action) {
   switch (action.type) {
     case MODAL_TOGGLE:
       return Object.assign({}, state, { open: action.payload });
